requirements.txt
streamlit
faiss-cpu
sentence-transformers
PyMuPDF
pandas
openpyxl
torch
transformers
pydub
audiorecorder
requests